Refactor game modules and redesign combat UI
