+++ src/app/app.js
@@ -1,0 +1,47 @@
+import { registerRenderer, setState, state, db } from './state.js';
+import { renderApp } from './ui/renderRoot.js';
+import { initBackground } from './background.js';
+import { describeGameState } from './game/core.js';
+
+export function setupApp(root) {
+  root.innerHTML = '';
+  const backgroundCanvas = document.createElement('canvas');
+  backgroundCanvas.id = 'bg-canvas';
+  root.appendChild(backgroundCanvas);
+
+  const uiLayer = document.createElement('div');
+  uiLayer.className = 'ui-layer';
+  root.appendChild(uiLayer);
+
+  registerRenderer(() => renderApp(uiLayer));
+  initBackground(backgroundCanvas);
+  renderApp(uiLayer);
+
+  db.subscribeAuth((authState) => {
+    if (authState.isLoading) {
+      setState({
+        auth: { loading: true, user: null, error: null },
+        screen: 'loading',
+      });
+      return;
+    }
+    if (authState.error) {
+      setState({
+        auth: { loading: false, user: null, error: authState.error.message },
+        screen: 'login',
+      });
+      return;
+    }
+    const user = authState.user ?? null;
+    setState({
+      auth: { loading: false, user, error: null },
+      screen: user ? 'menu' : 'login',
+    });
+  });
+
+  window.addEventListener('focus', () => {
+    if (state.screen === 'game') {
+      console.info('Current game state:', describeGameState());
+    }
+  });
+}
