+++ src/style.css
@@ -161,31 +161,66 @@ input {
 .game-view {
   position: relative;
   color: #fff;
-  gap: 0.8rem;
+  gap: 0.75rem;
 }
 
-.top-bar {
+.log-panel {
+  background: rgba(15, 18, 32, 0.82);
+  border-radius: 16px;
+  padding: 0.6rem 0.8rem;
+  box-shadow: inset 0 0 0 1px rgba(255, 255, 255, 0.04);
+  display: flex;
+  flex-direction: column;
+  gap: 0.5rem;
+}
+
+.log-header {
   display: flex;
   justify-content: space-between;
   align-items: center;
-  padding: 0.8rem 1rem;
-  border-radius: 14px;
-  background: rgba(15, 18, 32, 0.8);
-  box-shadow: inset 0 0 0 1px rgba(255, 255, 255, 0.05);
 }
 
-.mana-display {
-  font-weight: 700;
-  color: #ffd166;
+.log-recent,
+.log-dropdown ul {
+  margin: 0;
+  padding-left: 1.1rem;
+  font-size: 0.75rem;
+  line-height: 1.4;
 }
 
-.player-area {
-  background: rgba(12, 16, 28, 0.72);
+.log-recent li,
+.log-dropdown li {
+  margin-bottom: 0.35rem;
+}
+
+.log-dropdown {
+  max-height: 0;
+  overflow: hidden;
+  transition: max-height 0.25s ease;
+}
+
+.log-dropdown.open {
+  max-height: 180px;
+}
+
+.log-scroll {
+  max-height: 160px;
+  overflow-y: auto;
+  padding-right: 0.3rem;
+}
+
+.battlefield-area {
+  background: rgba(12, 16, 28, 0.75);
   border-radius: 16px;
-  padding: 0.6rem;
+  padding: 0.6rem 0.75rem;
   display: flex;
   flex-direction: column;
   gap: 0.6rem;
+}
+
+.battle-row {
+  display: flex;
+  flex-direction: column;
 }
 
 .player-header {
@@ -231,6 +266,19 @@ input {
   border-color: #ffd166;
 }
 
+.card.attacker-card {
+  border-color: #f97316;
+  box-shadow: 0 0 14px rgba(249, 115, 22, 0.4);
+}
+
+.card.blocker-selectable {
+  border-color: #38bdf8;
+}
+
+.card.targetable {
+  border-color: #c084fc;
+}
+
 .creature-card.summoning {
   opacity: 0.7;
 }
@@ -267,21 +315,32 @@ input {
   margin-top: 0.2rem;
 }
 
-.ability .mini {
+.mini {
   font-size: 0.65rem;
-  padding: 0.35rem 0.4rem;
+  padding: 0.35rem 0.5rem;
   background: rgba(255, 255, 255, 0.1);
-  width: 100%;
+  border-radius: 8px;
+  width: auto;
 }
 
 .hand-area {
   background: rgba(12, 16, 28, 0.8);
   border-radius: 16px;
   padding: 0.6rem;
+  display: flex;
+  flex-direction: column;
+  gap: 0.5rem;
 }
 
-.hand-area h3 {
-  margin: 0 0 0.4rem 0;
+.hand-header {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  margin-bottom: 0.2rem;
+}
+
+.hand-header h3 {
+  margin: 0;
 }
 
 .hand-cards {
@@ -290,59 +349,56 @@ input {
   gap: 0.5rem;
 }
 
-.log-area {
-  background: rgba(12, 16, 28, 0.65);
+.status-bar {
+  background: rgba(15, 18, 32, 0.78);
   border-radius: 16px;
-  padding: 0.6rem;
-  max-height: 140px;
-  overflow-y: auto;
-}
-
-.log-area ul {
-  margin: 0;
-  padding-left: 1.1rem;
-  font-size: 0.75rem;
-}
-
-.log-area li {
-  margin-bottom: 0.35rem;
-}
-
-.info-area {
-  background: rgba(12, 16, 28, 0.65);
-  border-radius: 16px;
-  padding: 0.6rem;
-  display: flex;
-  flex-direction: column;
-  gap: 0.5rem;
-}
-
-.life {
-  font-weight: 700;
-  font-size: 0.95rem;
-}
-
-.phase-controls {
+  padding: 0.6rem 0.8rem;
   display: flex;
   flex-direction: column;
   gap: 0.4rem;
 }
 
+.life-summary,
+.turn-summary {
+  display: flex;
+  justify-content: space-between;
+  font-size: 0.82rem;
+}
+
+.phase-controls {
+  display: flex;
+  flex-direction: column;
+  gap: 0.35rem;
+}
+
 .pending-overlay {
-  margin-top: 0.4rem;
-  background: rgba(30, 58, 138, 0.45);
+  background: rgba(255, 255, 255, 0.06);
   border-radius: 12px;
-  padding: 0.6rem;
+  padding: 0.55rem;
+  display: flex;
+  flex-direction: column;
+  gap: 0.4rem;
   font-size: 0.8rem;
 }
 
 .pending-overlay button {
-  margin-top: 0.4rem;
+  margin-top: 0.2rem;
 }
 
 .info {
   color: #93c5fd;
   font-size: 0.82rem;
+}
+
+.card[data-card],
+[data-location='hand'],
+[data-action] {
+  cursor: pointer;
+}
+
+button:disabled,
+button.disabled {
+  cursor: not-allowed;
 }
 
 @media (max-width: 540px) {
